+++ jest.config.js
@@ -6,5 +6,6 @@ module.exports = {
     "/dist/",
   ],
   testTimeout: 10000,
+  testMatch: ['**/*.test.ts'],
   testPathIgnorePatterns: ['integration'],
 };
